% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_modalities.R
\name{all_modalities}
\alias{all_modalities}
\title{Return All Modalities}
\usage{
all_modalities()
}
\value{
Vector of characters
}
\description{
Return the modalities for images where
packages were developed
}
