% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_ms_data.R
\name{copy_ms_data}
\alias{copy_ms_data}
\title{Copy MS Data to an output directory}
\usage{
copy_ms_data(outdir, ...)
}
\arguments{
\item{outdir}{Output directory for data}

\item{...}{Arguments to pass to \code{\link{get_image_filenames}}}
}
\value{
Logical if files are copied
}
\description{
Copies files from MS Lesion Package to an output directory
}
