% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{get_ids}
\alias{get_ids}
\alias{get_all_ids}
\alias{get_training_ids}
\alias{get_train_ids}
\alias{get_test_ids}
\title{Get IDs with Data in Package}
\usage{
get_ids(group = c("training", "test"))

get_all_ids()

get_training_ids()

get_train_ids()

get_test_ids()
}
\arguments{
\item{group}{group of IDs to gather.  If both \code{c("training", "test")},
all IDs are returned}
}
\value{
Vector of ids
}
\description{
Return the IDs for the people with data
}
