% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image_filenames.R
\name{get_image_filenames}
\alias{get_image_filenames}
\title{Get Image Filenames}
\usage{
get_image_filenames(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{get_image_filenames_df}}}
}
\description{
Return the filenames for the images
}
\examples{
get_image_filenames() 
}
