% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image_filenames.R
\name{get_image_filenames_df}
\alias{get_image_filenames_df}
\alias{get_image_filenames_matrix}
\alias{get_image_filenames_list}
\alias{get_image_filenames_list_by_subject}
\title{Get Image Filenames in a \code{data.frame}}
\usage{
get_image_filenames_df(group = c("training", "test"),
  modalities = all_modalities(), type = c("raw", "coregistered",
  "template"), derived = TRUE, long = TRUE)

get_image_filenames_matrix(...)

get_image_filenames_list(...)

get_image_filenames_list_by_subject(...)
}
\arguments{
\item{group}{group of IDs to gather.  If both 
\code{c("training", "test")},
all IDs are returned}

\item{modalities}{vector of image modalities within
\code{c("FLAIR", "T2", "T2", "PD")} to return}

\item{type}{type of data, either \code{"raw"}, \code{"coregistered"}
\code{"template"}}

\item{derived}{Get the derived images (tissue classes/brain mask)}

\item{long}{if \code{TRUE}, each row is a subject, 
visit, modality pair}

\item{...}{arguments passed to \code{\link{get_image_filenames_df}}}
}
\value{
Data.frame of filenames
}
\description{
Return a data.frame of filenames for the images
}
\examples{
get_image_filenames_df()
get_image_filenames_matrix()  
get_image_filenames_list()   
get_image_filenames_list_by_subject()     
}
