% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modality_df.R
\name{modality_df}
\alias{modality_df}
\title{All Modalities}
\usage{
modality_df()
}
\value{
\code{data.frame} of two columns:
\itemize{
\item{modality}: modality of image
}
}
\description{
Return the modalities for images
}
