% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_model.R
\docType{data}
\name{ms_model}
\alias{ms_model}
\title{Estimated Logistic Regression Model for MS}
\format{A \code{glm} object with estimates.}
\usage{
ms_model
}
\description{
A list containing the GLM 
for predicting MS lesions at a voxel level
}
\keyword{datasets}
